+++ src/include/LSP/Workspace.hpp
@@ -51,6 +51,7 @@ public:
     bool isIgnoredFile(const std::filesystem::path& path, const std::optional<ClientConfiguration>& givenConfig = std::nullopt);
 
     lsp::DocumentDiagnosticReport documentDiagnostics(const lsp::DocumentDiagnosticParams& params);
+    lsp::WorkspaceDiagnosticReport workspaceDiagnostics(const lsp::WorkspaceDiagnosticParams& params);
 
 private:
     void endAutocompletion(const lsp::CompletionParams& params);