+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added links to Luau Documentation for lint warnings
 - Added documentation to Enum types
 - Diagnostics of dependents (i.e., files which require a changed file) will now refresh to display type errors when the file changes
+- Added support for workspace-wide diagnostics to report in all files. You can enable this using `luau-lsp.diagnostics.workspace`. It is disabled by default
 
 ### Changed